voice_app/requirements.txt
fastapi
uvicorn[standard]
loguru
PyYAML
python-dotenv
openai
httpx 